+++ CHANGELOG.md
@@ -813,4 +813,4 @@
 - 增强 BotCtxUtil：
   - 对群聊私信进行区分并注解相关私信入口方法；
   - 新增 ChatId 获取方法；
-- 增强 QQAiClient - 添加从 BotCtx 中获取 ChatId 的重载方法。
+- 优化 QQAiClient - 优化方法参数列表和代码结构。